Add list, rename bot
